updated optimization of home page
